Move selectors/data database predicates into the databases module
getHasDatabaseWithJsonEngine and getHasDatabaseWithActionsEnabled are plain
predicates over database lists, so they move to
metabase/databases/utils/predicates. getHasOwnDatabase had no consumers
outside its own spec and is deleted as dead code. This also removes the one
module-boundary violation the selectors module carried (shared-utils
importing the shared-platform databases module).
