+++ frontend/src/metabase/common/components/NewItemMenu/NewItemMenu.tsx
@@ -1,7 +1,7 @@
 import { useListDatabasesQuery } from "metabase/api";
+import { getHasDatabaseWithJsonEngine } from "metabase/databases/utils/predicates";
 import { useDispatch, useSelector } from "metabase/redux";
 import { closeNavbar } from "metabase/redux/app";
-import { getHasDatabaseWithJsonEngine } from "metabase/selectors/data";
 import {
   canUserCreateNativeQueries,
   canUserCreateQueries,
